LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-form-submit-expected.txt



FAIL Navigating iframe loading='lazy' before it is loaded: form submit Error: assert_equals: expected "http://web-platform.test:8800/html/semantics/embedded-content/the-iframe-element/support/blank.htm?" but got "http://web-platform.test:8800/html/semantics/embedded-content/the-iframe-element/support/blank.htm?src"
